week6/v2/requirements.txt
streamlit>=1.30.0
openai>=1.9.0
Pillow>=10.0.0
python-dotenv>=1.0.0